backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTX==1.3.0
Flask-CORS==4.0.0
Flask-Caching==2.1.0
pandas==2.3.3
openpyxl==3.1.5
marshmallow==3.20.1
reportlab==4.2.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.0.1
